Fix slow mouse scrolling
Apparently setting mouse="" makes vim scroll the cursor instead of the
window, and when this is combined with :cursorcolumn true rendering
becomes very inefficient, even without any plugins. This is a workaround
that enables the mouse everywhere but then manually disables left and
right click to get similar behaviour as before.